Various chunk store / DynamoDB config refactorings.
(Motivation: tidying a few things up before handling #141)

- Move /dep.go into /pkg/ingester/client, where is belongs.
- Rename DynamoTableManager to TableManager, as there is nothing DynamoDB-specific about it - it would be useful to BigTable too.
- Rename DynamoTableClient interface to TableClient, as there is nothing DynamoDB-specici about it - it would be useful to BigTable too.
- Move dynamoTableClient (implementation of TableClient) into /pkg/chunk/table_manager.go, where it belongs.
- Delete chunk.DynamoTableClientConfig, as the table-manager-specific in-memory construction wasn't used.
- Move warning about unused DynamoDB table name in path to dynamoClientFromURL.
- Move duplicated OriginalTableName into PeriodicTableConfig, as per comment.
- Fix usage of flag.StringVar - should be f.StringVar.
